Stable - Setup complete
So far I’ve created a config for the database and a file to set it up
on the server and one query that calls all the data in the database and
stores it in a variable.

Next will be a GUI for displaying the results, and finally, a user
submission form that labels incoming results with “pending” status and
sends an email to the administrators, along with an admin GUI to
approve incoming requests and add their own.

Long road ahead.